+++ pkg/db/schema/0020_primary_keys_sampling.up.sql
@@ -38,11 +38,9 @@ ORDER BY (client_id, visitor_id, session_id, time)
 SAMPLE BY "visitor_id"
 SETTINGS index_granularity = 8192;
 
--- Run the following statements manually.
---INSERT INTO "event_new" SELECT * FROM "event" WHERE toDate(time) < today();
---RENAME TABLE "event" TO "event_backup";
---RENAME TABLE "event_new" TO "event";
---INSERT INTO "event" SELECT * FROM "event_backup" WHERE toDate(time) >= today();
-
--- If you're confident that everything is correct, drop the backup.
---DROP TABLE "event_backup";
+-- Consider running the following steps manually!
+INSERT INTO "event_new" SELECT * FROM "event" WHERE toDate(time) < today();
+RENAME TABLE "event" TO "event_backup";
+RENAME TABLE "event_new" TO "event";
+INSERT INTO "event" SELECT * FROM "event_backup" WHERE toDate(time) >= today();
+DROP TABLE "event_backup";